requirements.txt
Flask==0.10.1
Flask-API==0.6.2
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Pygments==2.0.2
REfO==0.13
SPARQLWrapper==1.6.4
Werkzeug==0.10.4
arrow==0.5.4
cffi==0.9.2
click==4.0
docopt==0.6.2
fancycompleter==0.4
gnureadline==6.3.3
html5lib==0.999
ipdb==0.8
ipython==3.1.0
isodate==0.5.1
itsdangerous==0.24
natto-py==0.4.0
ordereddict==1.1
pdbpp==0.8.2
prompt-toolkit==0.31
pycparser==2.10
pyparsing==2.0.3
pyrepl==0.8.4
python-dateutil==2.4.2
rdflib==4.2.0
redis==2.10.3
requests==2.7.0
rq==0.5.3
rq-dashboard==0.3.4
simplejson==3.7.3
six==1.9.0
times==0.7
ujson==1.33
wcwidth==0.1.4
wmctrl==0.1
wsgiref==0.1.2
